Move filter components to dedicated subfolder
