+++ src/Components/filters/Slider.test.js
